backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
geopy==2.4.1
pydantic==2.10.4
libsql-experimental
